requirements.txt
python-dotenv
pyttsx3
pyowm
langchain
langchain-groq
crewai
gTTS
pydub